agents/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
exceptiongroup==1.3.1
fastapi==0.136.1
h11==0.16.0
idna==3.13
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
requests==2.33.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
